+++ JrBoostLib/AbstractTrainer.h
@@ -10,8 +10,8 @@ protected:
 public:
     virtual ~AbstractTrainer() = default;
     virtual void setInData(CRefXXf inData) = 0;
-    virtual void setOutData(const ArrayXf& outData) = 0;
-    virtual void setWeights(const ArrayXf& weights) = 0;
+    virtual void setOutData(const ArrayXd& outData) = 0;
+    virtual void setWeights(const ArrayXd& weights) = 0;
     virtual void setOptions(const AbstractOptions& opt) = 0;
     virtual AbstractPredictor* train() const = 0;
